moves Task and subclasses, Parser, Storage, UI into packages as well as placed Chatterbox in a package
